+++ src/lib/prefs/AccentColorPref.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import Dropdown from '$lib/components/Dropdown.svelte';
+	import Info from '$lib/components/Info.svelte';
 	import Label from '$lib/components/Label.svelte';
 	import { colors, getAccentColor, setAccentColor } from '$lib/theme';
 	import { capitalize } from '$lib/util';
@@ -8,9 +9,9 @@
 </script>
 
 <div class="flex items-center">
-	<Label text="Accent color"
-		>Changes the color of highlighted elements, such as buttons and checkboxes.</Label
-	>
+	<Label>Accent color</Label>
+
+	<Info>Changes the color of highlighted elements, such as buttons and checkboxes.</Info>
 
 	<Dropdown
 		class="grow"