+++ src/lib/prefs/TogglePref.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
 	import Checkbox from '$lib/components/Checkbox.svelte';
+	import Info from '$lib/components/Info.svelte';
 	import Label from '$lib/components/Label.svelte';
-	import { invokeCommand } from '$lib/invoke';
 	import { confirm } from '@tauri-apps/plugin-dialog';
-	import { onMount } from 'svelte';
 
 	export let label: string;
 	export let disableMessage: string | null = null;
@@ -26,9 +25,13 @@
 </script>
 
 <div class="my-1 flex items-center">
-	<Label text={label}>
-		<slot />
+	<Label>
+		{label}
 	</Label>
+
+	<Info>
+		<slot />
+	</Info>
 
 	<Checkbox bind:value {onValueChanged} />
 </div>